requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
keyring==25.6.0
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.0
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
murmurhash==1.0.13
numpy==1.26.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
preshed==3.0.10
psycopg==3.2.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pync==2.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-jobspy @ git+https://github.com/Idoinnsmouth/JobSpy@d7415d6a5422802f23eba6d8ffbf1a45a303eac6
pytz==2025.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tls-client==1.0.1
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
